+++ database/migrations/1665657657397_catalog_products.ts
@@ -4,7 +4,7 @@ import CatalogProduct from 'App/Models/CatalogProduct'
 export default class extends BaseSchema {
   public override async up(): Promise<void> {
     this.schema.createTable(CatalogProduct.table, (table) => {
-      table.increments('id').primary()
+      table.bigInteger('id').primary()
       table.bigInteger('catalog_id').references('id').inTable('tbl_catalogs').onDelete('CASCADE')
       table.string('name', 50)
       table.decimal('price', 6, 2)
